fix(xhci): make the two u16 descriptor narrowings explicit for clang
Registering test_xhci_hid_composite pulled xhci_descparse.cpp into the
host-test build, where it is compiled under the host suite's own flags
rather than the kernel's. Both `lo | (hi << 8)` reads promote through
int, so assigning the result back to a u16 is an implicit narrowing that
clang rejects with -Werror,-Wimplicit-int-conversion.

GCC does not warn here, so my local host-test run (which uses g++) passed
137/137 while CI's two clang-18 jobs failed to compile. Verified this
time by reproducing both CI configurations exactly -- clang-18 with
DUETOS_HOST_TESTS_SANITIZERS and again with DUETOS_HOST_TESTS_TSAN --
each now 137/137.

The lesson generalises past this fix: linking a real kernel TU into a
host test subjects it to a second, stricter compiler, and the kernel
build passing says nothing about that.

+++ kernel/drivers/usb/xhci_descparse.cpp
@@ -44,7 +44,10 @@ void CaptureHidReportDescriptorLength(const u8* desc, u8 len, HidInterfaceRecord
     for (u8 i = 0; i < descriptorCount && off + 3 <= len; ++i, off = u8(off + 3))
     {
         const u8 descriptorType = desc[off];
-        const u16 descriptorLength = u16(desc[off + 1]) | (u16(desc[off + 2]) << 8);
+        // Both halves promote to int through the shift and the or, so the
+        // narrowing back to u16 has to be explicit: the host-test build
+        // compiles this TU under clang -Werror -Wimplicit-int-conversion.
+        const u16 descriptorLength = u16(u16(desc[off + 1]) | u16(u16(desc[off + 2]) << 8));
         if (descriptorType == kDescTypeReport && descriptorLength != 0)
         {
             iface.report_desc_length = descriptorLength;
@@ -412,7 +415,7 @@ bool ParseConfigForHidBoot(const u8* buf, u32 len, PortRecord& port)
         {
             const u8 bEndpointAddress = buf[off + 2];
             const u8 bmAttributes = buf[off + 3];
-            const u16 wMaxPacketSize = u16(buf[off + 4]) | (u16(buf[off + 5]) << 8);
+            const u16 wMaxPacketSize = u16(u16(buf[off + 4]) | u16(u16(buf[off + 5]) << 8));
             const u8 bInterval = buf[off + 6];
             if ((bmAttributes & kEpAttrTypeMask) == kEpAttrTypeInterrupt && (bEndpointAddress & kEpAddrDirIn))
             {
